fixed logo rendering error, dropdown to separate component
