+++ react-app/src/components/Dropdown/useDetectOutsideClick.js
