+++ react-app/src/components/NavBar/NavBar.css
@@ -7,10 +7,17 @@
     top: 0;
     left: 0;
     right: 0;
-    background-color: white;
+    background-color: 161315;
     z-index: 4;
     width: 100%;
     padding-bottom: 1px;
+}
+
+.header_logo {
+    width: 100px;
+    height: 100px;
+    overflow: hidden;
+    margin-top: -6px;
 }
 
 .header_left {
@@ -24,76 +31,6 @@
     align-items: center;
     padding-right: 10px;
 }
-
-.menu-container {
-    position: relative;
-}
-
-.menu {
-    background-color: #211C1F;
-    border-radius: 8px;
-    position: absolute;
-    top: 30px;
-    right: 0;
-    width: 300px;
-    box-shadow: 0 1px 8px rgba(0, 0, 0, 0.3);
-    opacity: 0;
-    visibility: hidden;
-    transform: translateY(-20px);
-  }
-
-  .menu.active {
-    opacity: 1;
-    visibility: visible;
-    transform: translateY(0);
-  }
-
-  .menu ul {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-  }
-
-  .menu li {
-    border-bottom: 2px solid #161315;
-  }
-
-  .menu li a {
-    text-decoration: none;
-    color: #449DD1;
-    padding: 15px 20px;
-    display: block;
-  }
-
-  .menu-trigger {
-    background: #ffffff;
-    border-radius: 90px;
-    cursor: pointer;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 4px 6px;
-    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.3);
-    border: none;
-    vertical-align: middle;
-    transition: box-shadow 0.4s ease;
-  }
-
-  .menu-trigger:hover {
-    box-shadow: 0 1px 8px rgba(0, 0, 0, 0.3);
-  }
-
-  .menu-trigger span {
-    font-weight: 700;
-    vertical-align: middle;
-    font-size: 14px;
-    margin: 0 10px;
-  }
-
-  .menu-trigger img {
-    border-radius: 90px;
-  }
-
 
 .home {
     padding-left: 10px;
